requirement.txt
autopep8==1.6.0
bidict==0.21.4
click==8.0.3
dnspython==2.2.0
eventlet==0.33.0
flake8==3.8.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
greenlet==1.1.2
https://github.com/benoitc/gunicorn/archive/ff58e0c6da83d5520916bc4cc109a529258d76e1.zip#egg=gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
kafka-python==2.0.2
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.6
pika==1.1.0
pycodestyle==2.8.0
pyflakes==2.2.0
python-engineio==4.3.1
python-socketio==5.5.1
six==1.16.0
toml==0.10.2
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
